feat: updated README and new tests
